Renamed project internals to Syzygy
