+++ source/syzygy/source/assets.cpp
