+++ source/syzygy/source/assets.hpp
