+++ source/syzygy/source/buffers.cpp
