+++ source/syzygy/source/buffers.hpp
