+++ source/syzygy/source/deferred/gbuffer.cpp
