+++ source/syzygy/source/deferred/gbuffer.hpp
