+++ source/syzygy/source/descriptors.cpp
