+++ source/syzygy/source/descriptors.hpp
