+++ source/syzygy/source/editor/editor.hpp
