+++ source/syzygy/source/editor/window.cpp
