+++ source/syzygy/source/editor/window.hpp
