+++ source/syzygy/source/engine.cpp
