+++ source/syzygy/source/engine.hpp
