+++ source/syzygy/source/enginetypes.hpp
