+++ source/syzygy/source/geometryhelpers.cpp
