+++ source/syzygy/source/geometryhelpers.hpp
