+++ source/syzygy/source/geometrystatics.hpp
