+++ source/syzygy/source/gputypes.hpp
