+++ source/syzygy/source/helpers.cpp
