+++ source/syzygy/source/helpers.hpp
