+++ source/syzygy/source/images.cpp
