+++ source/syzygy/source/images.hpp
