+++ source/syzygy/source/lights.hpp
