+++ source/syzygy/source/shadowpass.cpp
