+++ source/syzygy/source/shadowpass.hpp
