+++ source/syzygy/source/ui/propertytable.cpp
