+++ source/syzygy/source/ui/propertytable.hpp
